refactor: rename occurences of "rate" to "speed"
