+++ app/src/main/java/app/revanced/integrations/settings/SettingsEnum.java
@@ -18,8 +18,8 @@ public enum SettingsEnum {
     // Video settings
     OLD_STYLE_VIDEO_QUALITY_PLAYER_SETTINGS("revanced_use_old_style_quality_settings", true, ReturnType.BOOLEAN),
     REMEMBER_VIDEO_QUALITY_LAST_SELECTED("revanced_remember_video_quality_last_selected", true, ReturnType.BOOLEAN),
-    REMEMBER_PLAYBACK_RATE_LAST_SELECTED("revanced_remember_playback_rate_last_selected", true, ReturnType.BOOLEAN),
-    REMEMBER_PLAYBACK_RATE_LAST_SELECTED_VALUE("revanced_remember_playback_rate_last_selected_value", 1.0f, ReturnType.FLOAT),
+    REMEMBER_PLAYBACK_SPEED_LAST_SELECTED("revanced_remember_playback_rate_last_selected", true, ReturnType.BOOLEAN),
+    REMEMBER_PLAYBACK_SPEED_LAST_SELECTED_VALUE("revanced_remember_playback_speed_last_selected_value", 1.0f, ReturnType.FLOAT),
 
     // Whitelist settings
     //ToDo: Not used atm, Patch missing
